pyproject.toml
[project]
name = "multi-agent-research-assistant-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.128.0",
    "langchain>=1.2.0",
    "langchain-classic>=1.0.1",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.5",
    "langchain-openai>=1.1.6",
    "langchain-tavily>=0.2.16",
    "langgraph>=1.0.5",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "tavily-python>=0.7.17",
    "uvicorn>=0.40.0",
]
